requirements.txt
asgiref==3.8.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.11.12
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.2
daphne==4.0.0
Django==5.0.6
django-jazzmin==3.0.1
django-pwa==2.0.1
djangorestframework==3.15.2
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
packaging==24.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dotenv==1.0.1
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
zope.interface==8.0.1

